% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_oracle_report_sql_value.R
\name{sby_oracle_report_sql_value}
\alias{sby_oracle_report_sql_value}
\title{Escape SQL text literals}
\usage{
sby_oracle_report_sql_value(sql_value)
}
\arguments{
\item{sql_value}{A character value to escape for use in a SQL literal.}
}
\value{
A character scalar containing the formatted value or marker.
}
\description{
Duplicate single quotation marks in SQL character literals.
}
\keyword{internal}
