% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbyoracle-package.R
\docType{package}
\name{sbyoracle-package}
\alias{sbyoracle}
\alias{sbyoracle-package}
\title{R routines for Oracle Database 19c and later}
\description{
A growing library of Oracle database routines accessed through DBI. The initial
public interface generates comprehensive reports for tagged SQL statements.
Diagnostic helpers remain internal and are documented individually.
}
\seealso{
\code{\link[=sby_oracle_query_report]{sby_oracle_query_report()}}
}
\author{
\strong{Maintainer}: Joao B. G. de Brito \email{177784222+JoaoSabby@users.noreply.github.com}

Authors:
\itemize{
  \item Joao B. G. de Brito \email{177784222+JoaoSabby@users.noreply.github.com}
}

}
\keyword{internal}
